Add safe versions of CardTraitScalingUpgradeUnitXXX
The versions of these CardTraits that exist in the vanilla codebase aren't safe and will modify any upgrade the unit applies to itself (and others), since when applying a CardUpgrade from a unit considers the CardTraits the unit has.

This can be an issue with Enhancers that add a Trigger that plays CardEffectAddTempCardUpgradeToUnits like Heaven's Aid Revenge +1 attack buff.