Add IPostStartOfRunRelicEffect.cs
RelicEffectInterface for an artifact that runs its effects after a run has been successfully setup.

The use case for this is for ClassData#StarterRelics, If an artifact is given here it is added at the very very start of a run, before Covenant effects are ran. This is an issue since Covenant effects when ran will purge the initial starter deck and rebuild it from scratch. So if one wanted an artifact that works as Wurmtooth describes itself (Wurmtooth doesn't actually do anything) a Custom RelicEffect using IPostStartOfRunRelicEffect can be used to modify the starter deck without any additional patches.

This is probably why Wurmtooth doesn't do anything, and why the enable_corruption flag exists in the ClassData class.