Add correct versions of RoomStateDamageCostModifer and RoomStateHealCostModifer
These originals have a bug that allows a monster card that has a spawn effect that grants regen the bonus cost reduction.
And also checked if the cards have a permanent upgrade that damages or heals with an OnCast trigger

The logic of these have been changed to target cards with a damage/healing effect (can be extended) OR cards with a trigger type of OnCast with a damage/healing effect (from a temp or permanent upgrade).