Major Tweak for V1 CardUpgradeDataBuilder, IDs from Built CardUpgradeData will now generate a GUID. Support for not breaking existing RunHistory. This also makes the migration from Builders -> BuildersV2 seamless.
Minor Tweak for CustomClanHelper's FixShadowWarrior patches.

An issue that Enabling/Disabling CustomClanHelper will now break any Run History that includes the aforementioned unit.

So if ShadowWarrior's CardUpgrades are found, preemptively set the UpgradeTitle of its associated upgrades to what CustomClanHelper would set them to, that way even if CCH is removed (for some reason) (or installed later) the run history is unchanged.